requirements.txt
flask==2.2.3
flask-cors==3.0.10
google-generativeai==0.3.1
pdf2image==1.16.3
Pillow==9.5.0
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.28.2
google-cloud-storage==2.9.0
poppler-utils 